+++ src/py42/modules/detectionlists.py
@@ -1,3 +1,23 @@
+from py42.sdk.queries.query_filter import filter_attributes
+
+
+class RiskTags(object):
+    """The available RiskTags for :meth:`~py42.modules.detectionlists.DetectionListsModule.add_user_risk_tags()`
+    and :meth:`~py42.modules.detectionlists.DetectionListsModule.remove_user_risk_tags()`"""
+
+    FLIGHT_RISK = u"FLIGHT_RISK"
+    HIGH_IMPACT_EMPLOYEE = u"HIGH_IMPACT_EMPLOYEE"
+    ELEVATED_ACCESS_PRIVILEGES = u"ELEVATED_ACCESS_PRIVILEGES"
+    PERFORMANCE_CONCERNS = u"PERFORMANCE_CONCERNS"
+    SUSPICIOUS_SYSTEM_ACTIVITY = u"SUSPICIOUS_SYSTEM_ACTIVITY"
+    POOR_SECURITY_PRACTICES = u"POOR_SECURITY_PRACTICES"
+    CONTRACT_EMPLOYEE = u"CONTRACT_EMPLOYEE"
+
+    @staticmethod
+    def choices():
+        return filter_attributes(RiskTags)
+
+
 class DetectionListsModule(object):
     def __init__(self, microservice_client_factory):
         self._microservice_client_factory = microservice_client_factory
@@ -54,7 +74,7 @@ class DetectionListsModule(object):
         """Add or update notes related to the user.
 
         Args:
-            user_id (str or int): The Code42 userUid whose notes you want to update.
+            user_id (str or int): The userUid of the user whose notes you want to update.
             notes (str): User profile notes.
 
         Returns:
@@ -62,13 +82,13 @@ class DetectionListsModule(object):
         """
         factory = self._microservice_client_factory
         self._detection_list_user_client = factory.get_detection_list_user_client()
-        self._detection_list_user_client.update_notes(user_id, notes)
+        return self._detection_list_user_client.update_notes(user_id, notes)
 
     def add_user_risk_tags(self, user_id, tags):
         """Add one or more risk factor tags.
 
         Args:
-            user_id (str or int): The Code42 userUid whose risk factor tag(s) you want to update.
+            user_id (str or int): The userUid of the user whose risk factor tag(s) you want to update.
             tags (str or list of str ): A single tag or multiple tags in a list to be added. For
                 example: ``"tag1"`` or ``["tag1", "tag2"]``. For python version 2.X, pass ``u"str"``
                 instead of ``"str"``.
@@ -78,13 +98,13 @@ class DetectionListsModule(object):
         """
         factory = self._microservice_client_factory
         self._detection_list_user_client = factory.get_detection_list_user_client()
-        self._detection_list_user_client.add_risk_tags(user_id, tags)
+        return self._detection_list_user_client.add_risk_tags(user_id, tags)
 
     def remove_user_risk_tags(self, user_id, tags):
         """Remove one or more risk factor tags.
 
         Args:
-            user_id (str or int): The Code42 userUid whose risk factor tag(s) needs you want to remove.
+            user_id (str or int): The userUid of the user whose risk factor tag(s) needs you want to remove.
             tags (str or list of str ): A single tag or multiple tags in a list to be removed. For
                 example: ``"tag1"`` or ``["tag1", "tag2"]``. For python version 2.X, pass ``u"str"``
                 instead of ``"str"``.
@@ -94,13 +114,13 @@ class DetectionListsModule(object):
         """
         factory = self._microservice_client_factory
         self._detection_list_user_client = factory.get_detection_list_user_client()
-        self._detection_list_user_client.remove_risk_tags(user_id, tags)
+        return self._detection_list_user_client.remove_risk_tags(user_id, tags)
 
     def add_user_cloud_alias(self, user_id, alias):
         """Add a cloud alias to a user.
 
         Args:
-            user_id (str or int): The Code42 userUid whose alias you want to update.
+            user_id (str or int): The userUid of the user whose alias you want to update.
             alias (str): The alias to be added.
 
         Returns:
@@ -108,13 +128,13 @@ class DetectionListsModule(object):
         """
         factory = self._microservice_client_factory
         self._detection_list_user_client = factory.get_detection_list_user_client()
-        self._detection_list_user_client.add_cloud_alias(user_id, alias)
+        return self._detection_list_user_client.add_cloud_alias(user_id, alias)
 
     def remove_user_cloud_alias(self, user_id, alias):
         """Remove a cloud alias from a user.
 
         Args:
-            user_id (str or int): The user_id whose alias needs to be removed.
+            user_id (str or int): The userUid of the user whose alias needs to be removed.
             alias (str): The alias to be removed.
 
         Returns:
@@ -122,14 +142,14 @@ class DetectionListsModule(object):
         """
         factory = self._microservice_client_factory
         self._detection_list_user_client = factory.get_detection_list_user_client()
-        self._detection_list_user_client.remove_cloud_alias(user_id, alias)
+        return self._detection_list_user_client.remove_cloud_alias(user_id, alias)
 
     def refresh_user_scim_attributes(self, user_id):
         """Refresh SCIM attributes of a user.
         `REST documentation <https://ecm-default.prod.ffs.us2.code42.com/svc/swagger/index.html?urls.primaryName=v2#/User/UserControllerV2_RefreshUser>`__
 
         Args:
-            user_id (str or int): The Code42 userUid of the user whose attributes you wish to refresh.
+            user_id (str or int): The userUid of the user whose attributes you wish to refresh.
 
         Returns:
             :class:`py42.response.Py42Response`