+++ tests/clients/test_devices.py
@@ -94,3 +94,41 @@ class TestDeviceClient(object):
             pass
         py42.settings.items_per_page = 500
         assert mock_session.get.call_count == 3
+
+    def test_get_page_calls_get_with_expected_url_and_params(self, mock_session):
+        client = DeviceClient(mock_session)
+        client.get_page(20, True, True, "org", "user", "dest", True, True, 1000)
+        mock_session.get.assert_called_once_with(
+            "/api/Computer",
+            params={
+                "active": True,
+                "blocked": True,
+                "orgUid": "org",
+                "userUid": "user",
+                "targetComputerGuid": "dest",
+                "incBackupUsage": True,
+                "incCounts": True,
+                "pgNum": 20,
+                "pgSize": 1000,
+                "q": None,
+            },
+        )
+
+    def test_get_agent_state_calls_get_with_uri_and_params(
+        self, mock_session, successful_response
+    ):
+        mock_session.get.return_value = successful_response
+        client = DeviceClient(mock_session)
+        client.get_agent_state("DEVICE_ID", property_name="KEY")
+        expected_params = {"deviceGuid": "DEVICE_ID", "propertyName": "KEY"}
+        uri = u"/api/v14/agent-state/view-by-device-guid"
+        mock_session.get.assert_called_once_with(uri, params=expected_params)
+
+    def test_get_agent_full_disk_access_state_calls_get_agent_state_with_arguments(
+        self, mock_session, successful_response, mocker
+    ):
+        mock_session.get.return_value = successful_response
+        client = DeviceClient(mock_session)
+        client.get_agent_state = mocker.Mock()
+        client.get_agent_full_disk_access_state("DEVICE_ID")
+        client.get_agent_state.assert_called_once_with("DEVICE_ID", "fullDiskAccess")